+++ R/emmGrid-methods.R
@@ -585,6 +585,10 @@ update.emmGrid = function(object, ..., silent = FALSE) {
 #'   displayed is just enough to reasonably distinguish estimates from the ends
 #'   of their confidence intervals; but always at least 3 digits. If
 #'   \code{FALSE}, the system value \code{getOption("digits")} is used.}
+#' \item{\code{pval.digits}}{An integer indicating the precision with which
+#'   p-values are printed. Integers between \code{2} and \code{6} are acceptable; the default is 
+#' \code{4}. Floating point numbers are truncated. If a number outside the accepted range is
+#' used, the nearest acceptable integer will be used instead. Non-numeric arguments are ignored.}
 #' \item{\code{back.bias.adj}}{A logical value controlling whether we 
 #'   try to adjust bias when back-transforming. If \code{FALSE}, we use naive
 #'   back transformation. If \code{TRUE} \emph{and \code{sigma} is available and valid}, a
@@ -767,7 +771,8 @@ emm_defaults = list (
     estble.tol = 1e-8,        # tolerance for estimability checks
     simplify.names = TRUE,    # simplify names like data$x to just "x"
     back.bias.adj = FALSE,    # Try to bias-adjust back-transformations?
-    opt.digits = TRUE,        # optimize displayed digits?
+    opt.digits = TRUE,        # optimize displayed digits for estimates and CI's?
+    pval.digits = 4L,          # how many digits to list for p-values
     enable.submodel = TRUE,   # enable saving extra info for submodel
     rg.limit = 10000,         # limit on number of rows in a reference grid
     lmer.df = "kenward-roger",  # Use Kenward-Roger for df